initial net open loop controller

+++ src/Util.c
@@ -38,12 +38,12 @@ bool Util_StringEqual(const char* const a, const char* const b)
     return strcmp(a, b) == 0;
 }
 
-int64_t Util_Sqrt(const int64_t val)
+int32_t Util_Sqrt(const int64_t val)
 {
-    uint64_t a = 0;
-    uint64_t b = 0;
+    uint32_t a = 0;
+    uint32_t b = 0;
     if(val < 2)
-        return val;
+        return (int32_t) val;
     a = 1255;
     for(int32_t i = 0; i < 8; i++)
     {
